requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
anthropic==0.7.4
pytrends==4.9.2
pandas==2.0.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
email-validator==2.0.0
bcrypt==4.0.1 